+++ src/app/app-routing.module.ts
@@ -4,7 +4,11 @@ import {IdcardComponent} from "./idcard/idcard.component";
 import {PatientlistViewComponent} from "./patientlist-view/patientlist-view.component";
 import {ErrorComponent} from "./error/error.component";
 import {LogoutComponent} from "./logout/logout.component";
-import {canActivateAuthRole, canActivateChildAuthRole} from "./guards/auth-guard.service";
+import {
+  canActivateAuthRole,
+  canActivateChildAuthRole,
+  canActivateMlToken
+} from "./guards/auth-guard.service";
 import {Permission} from "./model/permission";
 import {CreatePatientComponent} from "./patient/create-patient/create-patient.component";
 import {EditPatientComponent} from "./patient/edit-patient/edit-patient.component";
@@ -20,6 +24,9 @@ import {
   BulkPseudonymizationComponent
 } from "./bulk-operations/bulk-pseudonymization/bulk-pseudonymization.component";
 import {
+  FailedAuthenticationComponent
+} from "./failed-authentication/failed-authentication.component";
+import {
   TentativeMatchesListComponent
 } from "./tentative-matches/tentative-matches-list/tentative-matches-list.component";
 import {
@@ -27,7 +34,7 @@ import {
 } from "./tentative-matches/solve-tentative-match/solve-tentative-match.component";
 
 export const routes: Routes = [
-  {
+  { // Note: access is only possible with OAuth authentication.
     path: '', canActivate: [canActivateAuthRole], canActivateChild: [canActivateChildAuthRole], children: [
       {path: '', pathMatch: 'full', redirectTo: 'patientlist'},
       {path: 'bulk-id-generation', component: BulkIdGenerationComponent, data : { permission: Permission.GENERATE_IDS, checkIdType:true}},
@@ -51,7 +58,13 @@ export const routes: Routes = [
 
     ]
   },
+  { // Access is only possible with a valid Mainzelliste token.
+    path: 'html', canActivate: [canActivateMlToken], canActivateChild: [canActivateChildAuthRole], children: [
+      {path: 'createPatient', component: CreatePatientComponent, data: { permission: Permission.CREATE_PATIENT }},
+    ]
+  },
   {path: 'access-denied', component: AccessDeniedComponent},
+  {path: 'auth-failed', component: FailedAuthenticationComponent},
   // Needs to be outside, because we want message why user couldn't authenticate
   {path: 'error', component: ErrorComponent},
   {path: 'logout', component: LogoutComponent},