+++ src/app/services/field.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Field} from "../model/field";
+import {Field, SemanticType} from "../model/field";
 import {AppConfigService} from "../app-config.service";
 import {GenderValue} from "../model/patientlist";
 
@@ -17,6 +17,13 @@ export class FieldService {
 
   getFields(): Array<Field> {
     return this.fields;
+  }
+
+  getSemanticFields(): { [key: string]: Field } {
+    return Object.fromEntries(this.fields.filter(f =>
+      f.semantic != undefined && f.semantic != SemanticType.UNDEFINED
+    )
+    .map(f => [f.semantic.valueOf(), f]));
   }
 
   getGenderFieldValues(): GenderValue[] {